more breaking - expose dae low level controls to dae module
