+++ manual_system_tests/dae_scan.py
@@ -14,18 +14,11 @@ from ophyd_async.plan_stubs import ensure_connected
 from ibex_bluesky_core.callbacks import ISISCallbacks
 from ibex_bluesky_core.devices.block import block_rw_rbv
 from ibex_bluesky_core.devices.simpledae import (
-    SimpleDae,
-    RunPerPointController,
     GoodFramesNormalizer,
     GoodFramesWaiter,
-)
-from ibex_bluesky_core.devices.simpledae.controllers import (
     RunPerPointController,
+    SimpleDae,
 )
-from ibex_bluesky_core.devices.simpledae.reducers import (
-    GoodFramesNormalizer,
-)
-from ibex_bluesky_core.devices.simpledae.waiters import GoodFramesWaiter
 from ibex_bluesky_core.fitting import Linear
 from ibex_bluesky_core.run_engine import get_run_engine
 from ibex_bluesky_core.utils import get_pv_prefix
