+++ src/ibex_bluesky_core/callbacks/__init__.py
@@ -20,9 +20,9 @@ from ibex_bluesky_core.callbacks._document_logger import DocLoggingCallback
 from ibex_bluesky_core.callbacks._file_logger import (
     HumanReadableFileCallback,
 )
+from ibex_bluesky_core.callbacks._fitting import LiveFit, LiveFitLogger
 from ibex_bluesky_core.callbacks._plotting import LivePlot, show_plot
 from ibex_bluesky_core.callbacks._utils import get_default_output_path
-from ibex_bluesky_core.callbacks.fitting import LiveFit, LiveFitLogger
 from ibex_bluesky_core.fitting import FitMethod
 
 logger = logging.getLogger(__name__)
