+++ src/ibex_bluesky_core/devices/__init__.py
@@ -11,7 +11,7 @@ from ophyd_async.epics.core import epics_signal_rw
 
 T = TypeVar("T", bound=SignalDatatype)
 
-__all__ = ["dehex_and_decompress", "compress_and_hex", "isis_epics_signal_rw"]
+__all__ = ["compress_and_hex", "dehex_and_decompress", "isis_epics_signal_rw"]
 
 
 def dehex_and_decompress(value: bytes) -> bytes:
