+++ src/ibex_bluesky_core/devices/dae/_period.py
