+++ src/ibex_bluesky_core/devices/simpledae/__init__.py
@@ -10,38 +10,46 @@ from ophyd_async.core import (
 )
 from typing_extensions import TypeVar
 
-from ibex_bluesky_core.devices.dae.dae import Dae
-from ibex_bluesky_core.devices.simpledae.controllers import (
+from ibex_bluesky_core.devices.dae import Dae
+from ibex_bluesky_core.devices.simpledae._controllers import (
     PeriodPerPointController,
     RunPerPointController,
 )
-from ibex_bluesky_core.devices.simpledae.reducers import (
-    MonitorNormalizer,
+from ibex_bluesky_core.devices.simpledae._reducers import (
     GoodFramesNormalizer,
+    MonitorNormalizer,
     PeriodGoodFramesNormalizer,
     ScalarNormalizer,
 )
-from ibex_bluesky_core.devices.simpledae.strategies import Controller, Reducer, Waiter
-from ibex_bluesky_core.devices.simpledae.waiters import (
+from ibex_bluesky_core.devices.simpledae._strategies import Controller, Reducer, Waiter
+from ibex_bluesky_core.devices.simpledae._waiters import (
     GoodFramesWaiter,
+    GoodUahWaiter,
+    MEventsWaiter,
     PeriodGoodFramesWaiter,
+    SimpleWaiter,
+    TimeWaiter,
 )
 from ibex_bluesky_core.utils import get_pv_prefix
 
 logger = logging.getLogger(__name__)
 
 __all__ = [
+    "GoodFramesNormalizer",
+    "GoodFramesWaiter",
+    "GoodUahWaiter",
+    "MEventsWaiter",
+    "MonitorNormalizer",
+    "PeriodGoodFramesNormalizer",
+    "PeriodGoodFramesWaiter",
+    "PeriodPerPointController",
+    "RunPerPointController",
+    "ScalarNormalizer",
     "SimpleDae",
+    "SimpleWaiter",
+    "TimeWaiter",
     "check_dae_strategies",
     "monitor_normalising_dae",
-    "RunPerPointController",
-    "PeriodPerPointController",
-    "GoodFramesWaiter",
-    "PeriodGoodFramesWaiter",
-    "MonitorNormalizer",
-    "GoodFramesNormalizer",
-    "PeriodGoodFramesNormalizer",
-    "ScalarNormalizer",
 ]
 
 TController_co = TypeVar("TController_co", bound="Controller", default="Controller", covariant=True)
