+++ src/ibex_bluesky_core/devices/simpledae/_controllers.py
@@ -10,9 +10,9 @@ from ophyd_async.core import (
     wait_for_value,
 )
 
-from ibex_bluesky_core.devices.dae.dae import RunstateEnum
-from ibex_bluesky_core.devices.dae.dae_controls import BeginRunExBits
-from ibex_bluesky_core.devices.simpledae.strategies import Controller
+from ibex_bluesky_core.devices.dae import RunstateEnum
+from ibex_bluesky_core.devices.dae._controls import BeginRunExBits
+from ibex_bluesky_core.devices.simpledae._strategies import Controller
 
 logger = logging.getLogger(__name__)
