+++ src/ibex_bluesky_core/devices/simpledae/_reducers.py
@@ -20,8 +20,8 @@ from ophyd_async.core import (
 )
 from scippneutron import conversion
 
-from ibex_bluesky_core.devices.dae.dae_spectra import DaeSpectra
-from ibex_bluesky_core.devices.simpledae.strategies import Reducer
+from ibex_bluesky_core.devices.dae._spectra import DaeSpectra
+from ibex_bluesky_core.devices.simpledae._strategies import Reducer
 
 logger = logging.getLogger(__name__)
