+++ src/ibex_bluesky_core/plans/reflectometry/_det_map_align.py
@@ -23,13 +23,13 @@ from ibex_bluesky_core.callbacks.reflectometry import (
     LivePColorMesh,
 )
 from ibex_bluesky_core.devices.simpledae import SimpleDae, check_dae_strategies
-from ibex_bluesky_core.devices.simpledae.controllers import (
+from ibex_bluesky_core.devices.simpledae._controllers import (
     PeriodPerPointController,
 )
-from ibex_bluesky_core.devices.simpledae.reducers import (
+from ibex_bluesky_core.devices.simpledae._reducers import (
     PeriodSpecIntegralsReducer,
 )
-from ibex_bluesky_core.devices.simpledae.strategies import Waiter
+from ibex_bluesky_core.devices.simpledae._strategies import Waiter
 from ibex_bluesky_core.fitting import Gaussian
 from ibex_bluesky_core.plan_stubs import call_qt_aware
