+++ src/ibex_bluesky_core/utils.py
@@ -7,7 +7,7 @@ from typing import Any, Protocol
 
 from bluesky.protocols import NamedMovable, Readable
 
-__all__ = ["centred_pixel", "get_pv_prefix", "NamedReadableAndMovable"]
+__all__ = ["NamedReadableAndMovable", "centred_pixel", "get_pv_prefix"]
 
 
 def centred_pixel(centre: int, pixel_range: int) -> list[int]:
