+++ tests/devices/simpledae/test_controllers.py
@@ -1,10 +1,10 @@
 import pytest
 from ophyd_async.testing import get_mock_put, set_mock_value
 
-from ibex_bluesky_core.devices.dae.dae import RunstateEnum
-from ibex_bluesky_core.devices.dae.dae_controls import BeginRunExBits
+from ibex_bluesky_core.devices.dae import RunstateEnum
+from ibex_bluesky_core.devices.dae._controls import BeginRunExBits
 from ibex_bluesky_core.devices.simpledae import SimpleDae
-from ibex_bluesky_core.devices.simpledae.controllers import (
+from ibex_bluesky_core.devices.simpledae._controllers import (
     PeriodPerPointController,
     RunPerPointController,
 )
