+++ tests/devices/simpledae/test_reducers.py
@@ -8,7 +8,7 @@ import scipp as sc
 from ophyd_async.testing import get_mock_put, set_mock_value
 
 from ibex_bluesky_core.devices.simpledae import SimpleDae
-from ibex_bluesky_core.devices.simpledae.reducers import (
+from ibex_bluesky_core.devices.simpledae._reducers import (
     GoodFramesNormalizer,
     MonitorNormalizer,
     PeriodGoodFramesNormalizer,
