+++ tests/devices/simpledae/test_waiters.py
@@ -4,7 +4,7 @@ import pytest
 from ophyd_async.testing import set_mock_value
 
 from ibex_bluesky_core.devices.simpledae import SimpleDae
-from ibex_bluesky_core.devices.simpledae.waiters import (
+from ibex_bluesky_core.devices.simpledae._waiters import (
     GoodFramesWaiter,
     GoodUahWaiter,
     MEventsWaiter,
