+++ tests/devices/test_dae.py
@@ -12,36 +12,36 @@ from bluesky.run_engine import RunEngine
 from ophyd_async.testing import get_mock_put, set_mock_value
 
 from ibex_bluesky_core.devices import compress_and_hex, dehex_and_decompress
-from ibex_bluesky_core.devices.dae.dae import Dae, RunstateEnum
-from ibex_bluesky_core.devices.dae.dae_period_settings import (
+from ibex_bluesky_core.devices.dae import Dae, RunstateEnum
+from ibex_bluesky_core.devices.dae._helpers import (
+    _convert_xml_to_names_and_values,
+    _set_value_in_dae_xml,
+)
+from ibex_bluesky_core.devices.dae._period_settings import (
     DaePeriodSettings,
     DaePeriodSettingsData,
     PeriodSource,
     PeriodType,
     SinglePeriodSettings,
 )
-from ibex_bluesky_core.devices.dae.dae_settings import (
+from ibex_bluesky_core.devices.dae._settings import (
     DaeSettings,
     DaeSettingsData,
     DaeTimingSource,
 )
-from ibex_bluesky_core.devices.dae.dae_spectra import VARIANCE_ADDITION, DaeSpectra
-from ibex_bluesky_core.devices.dae.dae_tcb_settings import (
-    TCBCalculationMethod,
+from ibex_bluesky_core.devices.dae._spectra import VARIANCE_ADDITION, DaeSpectra
+from ibex_bluesky_core.devices.dae._tcb_settings import (
     DaeTCBSettings,
     DaeTCBSettingsData,
+    TCBCalculationMethod,
+    TCBTimeUnit,
     TimeRegime,
     TimeRegimeMode,
     TimeRegimeRow,
-    TCBTimeUnit,
 )
-from ibex_bluesky_core.devices.dae._helpers import (
-    _convert_xml_to_names_and_values,
-    _set_value_in_dae_xml,
-)
-from src.ibex_bluesky_core.devices.dae.dae_controls import BeginRunExBits
-from src.ibex_bluesky_core.devices.dae.dae_period_settings import _convert_period_settings_to_xml
-from src.ibex_bluesky_core.devices.dae.dae_tcb_settings import _convert_tcb_settings_to_xml
+from src.ibex_bluesky_core.devices.dae._controls import BeginRunExBits
+from src.ibex_bluesky_core.devices.dae._period_settings import _convert_period_settings_to_xml
+from src.ibex_bluesky_core.devices.dae._tcb_settings import _convert_tcb_settings_to_xml
 from tests.conftest import MOCK_PREFIX
 from tests.devices.dae_testing_data import (
     dae_settings_template,
