+++ tests/plans/test_det_map_align.py
@@ -7,8 +7,8 @@ from ophyd_async.core import soft_signal_rw
 from ophyd_async.testing import callback_on_mock_put, set_mock_value
 
 from ibex_bluesky_core.devices.simpledae import PeriodPerPointController, SimpleDae
-from ibex_bluesky_core.devices.simpledae.reducers import PeriodSpecIntegralsReducer
-from ibex_bluesky_core.devices.simpledae.strategies import Waiter
+from ibex_bluesky_core.devices.simpledae._reducers import PeriodSpecIntegralsReducer
+from ibex_bluesky_core.devices.simpledae._strategies import Waiter
 from ibex_bluesky_core.plans.reflectometry import (
     angle_scan_plan,
     height_and_angle_scan_plan,
